Switch screenshots to check for shutdown

+++ src/screenshots.rs
@@ -1,7 +1,8 @@
-use anyhow::Result;
+use anyhow::{anyhow, Result};
 use log::{error, info};
 use reqwest::Body;
 use sha1::{Digest, Sha1};
+use std::collections::VecDeque;
 use std::path::{Path, PathBuf};
 use std::time::Duration;
 use tokio::fs::{remove_file, File};
@@ -23,6 +24,7 @@ pub struct Screenshots {
     rx: mpsc::UnboundedReceiver<Event>,
     screenshot_url: String,
     extra_directory: String,
+    buffer: VecDeque<Event>,
 }
 
 pub fn launch() -> (ScreenshotsPre, mpsc::UnboundedSender<Event>) {
@@ -36,6 +38,7 @@ impl ScreenshotsPre {
             rx: self.rx,
             screenshot_url,
             extra_directory,
+            buffer: VecDeque::new(),
         };
         screenshots.start().await
     }
@@ -43,25 +46,62 @@ impl ScreenshotsPre {
 
 impl Screenshots {
     pub async fn start(mut self) -> Result<()> {
-        while let Some(event) = self.rx.recv().await {
-            info!("Handling event {event:?}");
-            match event {
-                Event::UploadScreenshot(path, directory) => {
-                    self.upload_path_to_directory(&path, &directory).await;
-                    if let Err(e) = remove_file(&path).await {
-                        error!("Could not remove uploaded screenshot file {path:?}: {e:?}");
-                    }
+        loop {
+            // if we have a buffer, then we want to just check on the channel and continue
+            // otherwise block
+            let event = if self.buffer.is_empty() {
+                self.rx.recv().await
+            } else {
+                match self.rx.try_recv() {
+                    Ok(e) => Some(e),
+                    Err(tokio::sync::mpsc::error::TryRecvError::Disconnected) => break,
+                    _ => None,
                 }
+            };
 
-                Event::UploadExtra(path) => {
-                    self.upload_path_to_directory(&path, &self.extra_directory)
-                        .await;
-                    if let Err(e) = remove_file(&path).await {
-                        error!("Could not remove extra screenshot file {path:?}: {e:?}");
-                    }
+            if let Some(event) = event {
+                info!("Handling event {event:?}");
+                match event {
+                    Event::StartShutdown => break,
+                    _ => self.buffer.push_back(event),
                 }
+            }
 
-                Event::StartShutdown => break,
+            if let Some(event) = self.buffer.pop_front() {
+                match &event {
+                    Event::UploadScreenshot(path, directory) => {
+                        if let Err(e) = self.upload_path_to_directory(path, directory).await {
+                            error!("Could not upload {path:?}: {e:?}");
+                            self.buffer.push_front(event);
+                            info!("Sleeping for 5s before trying again");
+                            tokio::time::sleep(tokio::time::Duration::from_secs(5)).await;
+                            continue;
+                        }
+
+                        if let Err(e) = remove_file(&path).await {
+                            error!("Could not remove uploaded screenshot file {path:?}: {e:?}");
+                        }
+                    }
+
+                    Event::UploadExtra(path) => {
+                        if let Err(e) = self
+                            .upload_path_to_directory(path, &self.extra_directory)
+                            .await
+                        {
+                            error!("Could not upload {path:?}: {e:?}");
+                            self.buffer.push_front(event);
+                            info!("Sleeping for 5s before trying again");
+                            tokio::time::sleep(tokio::time::Duration::from_secs(5)).await;
+                            continue;
+                        }
+
+                        if let Err(e) = remove_file(&path).await {
+                            error!("Could not remove extra screenshot file {path:?}: {e:?}");
+                        }
+                    }
+
+                    Event::StartShutdown => unreachable!(),
+                }
             }
         }
 
@@ -94,7 +134,7 @@ impl Screenshots {
         }
     }
 
-    async fn upload_path_to_directory(&self, path: &Path, directory: &str) {
+    async fn upload_path_to_directory(&self, path: &Path, directory: &str) -> Result<()> {
         let mut url = format!("{}/{directory}", self.screenshot_url);
         let extension = path
             .extension()
@@ -111,53 +151,30 @@ impl Screenshots {
             url.push_str(&param);
         }
 
-        loop {
-            match File::open(path).await {
-                Ok(file) => {
-                    let stream = FramedRead::new(file, BytesCodec::new());
-                    let body = Body::wrap_stream(stream);
+        let file = File::open(path).await?;
+        let stream = FramedRead::new(file, BytesCodec::new());
+        let body = Body::wrap_stream(stream);
 
-                    let builder = reqwest::ClientBuilder::new().timeout(Duration::from_secs(30));
-                    let client = builder.build().unwrap();
-                    match client
-                        .post(&url)
-                        .header(reqwest::header::CONTENT_TYPE, content_type)
-                        .body(body)
-                        .send()
-                        .await
-                    {
-                        Ok(res) => {
-                            if res.status().is_success() {
-                                match res.text().await {
-                                    Ok(message) => {
-                                        info!(
-                                            "Successfully uploaded {path:?} to {url:?}: {message}"
-                                        );
-                                        break;
-                                    }
-                                    Err(e) => {
-                                        error!("Failed to parse upload response for {path:?} using {url:?} into text: {e:?}");
-                                    }
-                                }
-                            } else {
-                                error!(
-                                    "Failed to upload {path:?} using {url:?}: got status code {}",
-                                    res.status()
-                                );
-                            }
-                        }
-                        Err(e) => {
-                            error!("Failed to upload {path:?} using {url:?}: {e:?}");
-                        }
-                    }
-                }
-                Err(e) => {
-                    error!("Could not open {path:?} for reading: {e:?}");
-                }
-            }
+        let builder = reqwest::ClientBuilder::new().timeout(Duration::from_secs(30));
+        let client = builder.build()?;
 
-            info!("Sleeping for 5s before trying again");
-            tokio::time::sleep(tokio::time::Duration::from_secs(5)).await;
+        let res = client
+            .post(&url)
+            .header(reqwest::header::CONTENT_TYPE, content_type)
+            .body(body)
+            .send()
+            .await?;
+
+        if !res.status().is_success() {
+            return Err(anyhow!(
+                "Failed to upload {path:?} using {url:?}: got status code {}",
+                res.status()
+            ));
         }
+
+        let message = res.text().await?;
+        info!("Successfully uploaded {path:?} to {url:?}: {message}");
+
+        Ok(())
     }
 }
